moved all_estimators to registry
