+++ sktime/forecasting/tests/test_all_forecasters.py
@@ -37,7 +37,7 @@ from sktime.forecasting.tests._config import VALID_INDEX_FH_COMBINATIONS
 from sktime.performance_metrics.forecasting import (
     mean_absolute_percentage_error,
 )
-from sktime.utils import all_estimators
+from sktime.registry import all_estimators
 from sktime.utils._testing.estimator_checks import _construct_instance
 from sktime.utils._testing.forecasting import _assert_correct_pred_time_index
 from sktime.utils._testing.forecasting import _get_expected_index_for_update_predict
