+++ sktime/registry/README.md
@@ -5,3 +5,6 @@ This folder contains registries of key `sktime` concepts and objects:
 * `_tags` : estimator tags used within `sktime`
 
 New objects should be added to this registry for automated recognition by the test suite.
+
+The `_lookup` sub-module contains runtime inspection functionality for the registry:
+* `all_estimators` for returning lists of estimators by specified selection criteria
