+++ sktime/registry/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from sktime.registry._tags import ESTIMATOR_TAG_REGISTER, ESTIMATOR_TAG_LIST
 from sktime.registry._base_classes import (
     BASE_CLASS_REGISTER,
     BASE_CLASS_LIST,
@@ -12,7 +11,18 @@ from sktime.registry._base_classes import (
     TRANSFORMER_MIXIN_SCITYPE_LIST,
 )
 
+from sktime.registry._lookup import (
+    all_estimators,
+    _get_name,
+    _has_tag,
+)
+
+from sktime.registry._tags import ESTIMATOR_TAG_REGISTER, ESTIMATOR_TAG_LIST
+
 __all__ = [
+    "all_estimators",
+    "_get_name",
+    "_has_tag",
     "ESTIMATOR_TAG_LIST",
     "ESTIMATOR_TAG_REGISTER",
     "BASE_CLASS_REGISTER",
