+++ sktime/transformations/tests/test_all_transformers.py
@@ -14,8 +14,8 @@ from sktime.transformations.base import _PanelToPanelTransformer
 from sktime.transformations.base import _PanelToTabularTransformer
 from sktime.transformations.base import _SeriesToPrimitivesTransformer
 from sktime.transformations.base import _SeriesToSeriesTransformer
-from sktime.utils import all_estimators
-from sktime.utils import _has_tag
+from sktime.registry import all_estimators
+from sktime.registry import _has_tag
 from sktime.utils._testing.estimator_checks import _assert_array_almost_equal
 from sktime.utils._testing.estimator_checks import _construct_instance
 from sktime.utils._testing.estimator_checks import _make_args
